refactor(preset-mini,preset-wind): move variants to preset-wind
- move `motion-reduce`/`motion-safe`
- move `portrait`/`landscape`
