+++ packages/preset-mini/src/variants/default.ts
@@ -6,7 +6,7 @@ import { variantCombinators } from './combinators'
 import { variantColorsMediaOrClass } from './dark'
 import { variantLanguageDirections } from './directions'
 import { variantImportant, variantLayer, variantNegative, variantScope } from './misc'
-import { variantCustomMedia, variantMotions, variantOrientations, variantPrint } from './media'
+import { variantCustomMedia, variantPrint } from './media'
 import { partClasses, variantPseudoClassFunctions, variantPseudoClasses, variantPseudoElements, variantTaggedPseudoClasses } from './pseudo'
 
 export const variants = (options: PresetMiniOptions): Variant<Theme>[] => [
@@ -15,8 +15,6 @@ export const variants = (options: PresetMiniOptions): Variant<Theme>[] => [
   variantImportant,
   variantPrint,
   variantCustomMedia,
-  ...variantOrientations,
-  ...variantMotions,
   variantBreakpoints,
   ...variantCombinators,
   variantPseudoClasses,
