+++ packages/preset-wind/src/variants/default.ts
@@ -3,6 +3,7 @@ import { variants as miniVariants } from '@unocss/preset-mini/variants'
 import type { Theme, UnoOptions } from '..'
 import { variantCombinators } from './combinators'
 import { variantColorsScheme } from './dark'
+import { variantMotions, variantOrientations } from './media'
 import { variantSpaceAndDivide } from './misc'
 import { placeholderModifier } from './placeholder'
 
@@ -10,6 +11,8 @@ export const variants = (options: UnoOptions): Variant<Theme>[] => [
   placeholderModifier,
   variantSpaceAndDivide,
   ...miniVariants(options),
+  ...variantOrientations,
+  ...variantMotions,
   ...variantCombinators,
   ...variantColorsScheme,
 ]
