+++ packages/preset-wind/src/variants/index.ts
@@ -2,5 +2,6 @@
 export * from './combinators'
 export * from './dark'
 export * from './default'
+export * from './media'
 export * from './misc'
 export * from './placeholder'
